+++ packages/ui/Cancelable/cancelHandler.js
@@ -1,4 +1,4 @@
-import {forKeyCode} from '@enact/core/handle';
+import {forKey} from '@enact/core/handle';
 
 /**
  * Array of cancel handlers. If any of these returns `false`, `forCancel` will return `false`;
@@ -8,7 +8,7 @@ import {forKeyCode} from '@enact/core/handle';
  */
 const cancelHandlers = [
 	// always support the Escape key
-	forKeyCode(27)
+	forKey('cancel')
 ];
 
 /**